+++ ichnaea/data/tasks.py
@@ -56,20 +56,20 @@ def insert_measures(self, items=None, email=None, ip=None, nickname=None,
 
 
 @celery_app.task(base=BaseTask, bind=True, queue='celery_insert')
-def insert_measures_cell(self, entries, userid=None, utcnow=None):
+def insert_measures_cell(self, entries, userid=None):
     with self.redis_pipeline() as pipe:
         with self.db_session() as session:
-            queue = CellObservationQueue(self, session, pipe, utcnow=utcnow)
-            length = queue.insert(entries, userid=userid)
+            queue = CellObservationQueue(self, session, pipe)
+            length = queue.insert(entries)
     return length
 
 
 @celery_app.task(base=BaseTask, bind=True, queue='celery_insert')
-def insert_measures_wifi(self, entries, userid=None, utcnow=None):
+def insert_measures_wifi(self, entries, userid=None):
     with self.redis_pipeline() as pipe:
         with self.db_session() as session:
-            queue = WifiObservationQueue(self, session, pipe, utcnow=utcnow)
-            length = queue.insert(entries, userid=userid)
+            queue = WifiObservationQueue(self, session, pipe)
+            length = queue.insert(entries)
     return length